# Adjust propagation kernel launch bounds