--- a/device/cuda/src/fitting/kernels/specializations/fit_backward_src.cuh
+++ b/device/cuda/src/fitting/kernels/specializations/fit_backward_src.cuh
@@ -14,7 +14,7 @@
 namespace traccc::cuda {
 namespace kernels {
 template <typename fitter_t>
-__global__ __launch_bounds__(128, 8) void fit_backward(
+__global__ __launch_bounds__(128) void fit_backward(
     const fitting_config cfg, const device::fit_payload<fitter_t> payload) {
     device::fit_backward<fitter_t>(details::global_index1(), cfg, payload);
 }