This commit modifies the launch bounds for our propagation and fitting kernels, removing the minimum number of blocks per SM. This helps performance as it allows the kernel to use more registers. The parallelism on these kernels is currently bound by local memory, anyway.