@stephenswat 2025-08-19
Before:

```
16:13:38    ThroughputExample             INFO      Throughput:            Warm-up processing  40.5623 ms/event, 24.6535 events/s
16:13:38    ThroughputExample             INFO                    Event processing  35.6418 ms/event, 28.057 events/s
```

After:

```
16:19:29    ThroughputExample             INFO      Throughput:            Warm-up processing  34.2732 ms/event, 29.1773 events/s
16:19:29    ThroughputExample             INFO                    Event processing  28.4828 ms/event, 35.1089 events/s
```